Move preconditions to src/common/base, add test
