+++ src/planning/test/recipe-index-test.ts
@@ -13,7 +13,7 @@ import {Arc} from '../../runtime/arc.js';
 import {Loader} from '../../runtime/loader.js';
 import {Manifest} from '../../runtime/manifest.js';
 import {MockSlotComposer} from '../../runtime/testing/mock-slot-composer.js';
-import {checkDefined} from '../../runtime/testing/preconditions.js';
+import {checkDefined} from '../../common/base/preconditions.js';
 import {RecipeIndex} from '../recipe-index.js';
 import {Id, ArcId} from '../../runtime/id.js';
